+++ src/button/button.jsx
@@ -8,15 +8,12 @@ import Type from 'prop-types';
 
 import cn from '../cn';
 import keyboardCode from '../lib/keyboard-code';
-import { deprecated } from '../lib/prop-types';
-import performance from '../performance';
 
 /**
  * Компонент кнопки (да, она нажимается!).
  */
 @cn('button')
-@performance()
-class Button extends React.Component {
+class Button extends React.PureComponent {
     static propTypes = {
         /** Текст кнопки */
         text: Type.node,
@@ -27,7 +24,7 @@ class Button extends React.Component {
         /** Список произвольных элементов в правом слоте */
         leftAddons: Type.node,
         /** Тип кнопки */
-        view: Type.oneOf(['default', 'action', 'extra']),
+        view: Type.oneOf(['default', 'action', 'extra', 'rounded']),
         /** Поведение кнопки */
         type: Type.oneOf(['button', 'reset', 'submit']),
         /** HTML элемент, которым будет компонент в DOM */
@@ -41,7 +38,7 @@ class Button extends React.Component {
         /** Отображение кнопки в состоянии фокуса */
         focused: Type.bool,
         /** Псевдо представление кнопки */
-        pseudo: deprecated(Type.bool, 'Pseudo buttons are deprecated, remove this prop'),
+        pseudo: Type.bool,
         /** Идентификатор компонента в DOM */
         id: Type.string,
         /** Отключает валидацию полей формы, у которых есть атрибут pattern */
@@ -142,11 +139,12 @@ class Button extends React.Component {
     }
 
     render(cn) {
-        let buttonElement = this.props.tag === 'span' ? 'span' : 'button';
-        const isButton = buttonElement === 'button';
+        const isButton = this.props.tag !== 'span';
 
         let buttonProps = {
-            ref: (control) => { this.control = control; },
+            ref: (control) => {
+                this.control = control;
+            },
             role: 'button',
             id: this.props.id,
             name: this.props.name,
@@ -180,27 +178,36 @@ class Button extends React.Component {
         };
 
         let buttonContent = [
-            this.props.leftAddons &&
-                <span key='left-addons'>
+            this.props.leftAddons && (
+                <span key='left-addons' className={ cn('addon') }>
                     { this.props.leftAddons }
-                </span>,
-            this.props.icon &&
-                <span key='icon' className={ cn('icon') }>
-                    { this.props.icon }
-                </span>,
-            (this.props.children || this.props.text) &&
-                <span key='text' className={ cn('text') }>
-                    { this.props.children || this.props.text }
-                </span>,
-            this.props.rightAddons &&
-                <span key='right-addons'>
+                </span>
+            ),
+            (this.props.children || this.props.text || this.props.icon) && (
+                <span key='content' className={ cn('content') }>
+                    { this.props.icon && (
+                        <span key='icon' className={ cn('icon') }>
+                            { this.props.icon }
+                        </span>
+                    ) }
+                    { (this.props.children || this.props.text) && (
+                        <span key='text' className={ cn('text') }>
+                            { this.props.children || this.props.text }
+                        </span>
+                    ) }
+                </span>
+            ),
+            this.props.rightAddons && (
+                <span key='right-addons' className={ cn('addon') }>
                     { this.props.rightAddons }
                 </span>
+            )
         ];
 
-        return React.createElement(buttonElement,
-            buttonProps,
-            buttonContent
+        return isButton ? (
+            <button { ...buttonProps }>{ buttonContent }</button>
+        ) : (
+            <span { ...buttonProps }>{ buttonContent }</span>
         );
     }