+++ src/intl-phone-input/intl-phone-input.jsx
@@ -10,7 +10,6 @@ import Input from '../input/input';
 import Select from '../select/select';
 
 import cn from '../cn';
-import performance from '../performance';
 
 import countries from '../lib/countries';
 import getRelatedTarget from '../lib/related-target';
@@ -22,8 +21,7 @@ const MAX_DIAL_CODE_LENGTH = 4;
  *
  */
 @cn('intl-phone-input', Input, Select)
-@performance()
-class IntlPhoneInput extends React.Component {
+class IntlPhoneInput extends React.PureComponent {
     static propTypes = Input.propTypes;
 
     static defaultProps = {