+++ src/paragraph/paragraph.jsx
@@ -6,14 +6,12 @@ import React from 'react';
 import Type from 'prop-types';
 
 import cn from '../cn';
-import performance from '../performance';
 
 /**
  * Компонент параграфа текста.
  */
 @cn('paragraph')
-@performance()
-class Paragraph extends React.Component {
+class Paragraph extends React.PureComponent {
     static propTypes = {
         /** Тип параграфа */
         view: Type.oneOf(['lead', 'normal']),