+++ src/swipeable/swipeable.jsx
@@ -3,8 +3,6 @@ import Type from 'prop-types';
 
 import autobind from 'core-decorators/lib/autobind';
 
-import performance from '../performance';
-
 export const getCoordinates = ({
     touches, changedTouches, clientX, clientY
 }) => (
@@ -13,8 +11,7 @@ export const getCoordinates = ({
         : { clientX, clientY }
 );
 
-@performance()
-export default class Swipeable extends React.Component {
+export default class Swipeable extends React.PureComponent {
     static propTypes = {
         /** Число пикселей, на которое нужно сместиться, чтобы запустить функцию по свайпу */
         delta: Type.number,